# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smart Airport Ride Pooling API
  version: 1.0.0
  description: |
    Production-grade backend for intelligent airport ride pooling.
    Pools passengers into shared cabs with route optimization,
    detour limits, luggage capacity, concurrency safety, and dynamic pricing.
  contact:
    name: Airport Ride Pooling Team
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
  /ride/request:
    post:
      tags: [Ride]
      summary: Create a ride request
      description: Creates a new ride request and enqueues it for async pool matching via BullMQ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [passengerId, destLat, destLng, destAddress]
              properties:
                passengerId:
                  type: string
                  format: uuid
                  description: UUID of the passenger
                destLat:
                  type: number
                  format: float
                  example: 13.0827
                  description: Destination latitude
                destLng:
                  type: number
                  format: float
                  example: 80.2707
                  description: Destination longitude
                destAddress:
                  type: string
                  example: "T. Nagar, Chennai"
                  description: Human-readable destination
                luggageCount:
                  type: integer
                  default: 1
                  minimum: 0
                  maximum: 10
                maxDetourRatio:
                  type: number
                  format: float
                  default: 1.4
                  minimum: 1.0
                  maximum: 3.0
                  description: Max acceptable detour ratio (1.4 = 40% extra distance)
      responses:
        '201':
          description: Ride request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  rideRequestId:
                    type: string
                    format: uuid
                  jobId:
                    type: string
                  status:
                    type: string
                    enum: [PENDING]
        '400':
          description: Validation error
        '404':
          description: Passenger not found
        '429':
          description: Rate limited
...